Add new version of web client using UnityWebRequest.

+++ vFrame.Core.Unity/Download/DownloadAgentBase.cs
@@ -14,7 +14,7 @@ namespace vFrame.Core.Download
 
         public bool TaskDone { get; private set; }
 
-        public float Timeout { get; set; }
+        public int Timeout { get; set; }
 
         public float ProgressUpdateInterval { get; set; }
